use unique service name
On acm.xidian.edu.cn, we are getting something like

    2022-11-25 16:56:49
      ERROR: judgement failed: systemd error:
      cannot start the transient service:
      org.freedesktop.systemd1.UnitExists:
      Unit opoj-runner-7.service already exists.
    2022-11-25 16:56:49
      INFO: verdict = JudgementFailed

This is because we are relying on systemd to collect the transient
services.  Then there is a race condition: if systemd collects the
service later than we attempt to start the next run, bad thing will
happen.

Ideally we should make sure systemd has already collected the previous
service, then start the next one.  But it's not a trivial job, so for
now just use an unique name for the transient services.